Major scafolding

+++ FintrakBanking.Repositories/Credit/LoanApplicationRepository.cs
@@ -240,7 +240,7 @@ namespace FintrakBanking.Repositories.Credit
                 LoanTypeId = loan.loanTypeId,
                 LoanStatusId = loanStatusId,
                 CompanyId = loan.companyId,
-                BranchId = loan.branchId,
+                BranchId = (short)loan.branchId,
                 Tenor = loan.tenor, 
                 RelationshipOfficerId = loan.relationshipOfficerId,
                 RelationshipManagerId = loan.relationshipManagerId,
